Latest version objectify works and integrated masons part
objectify is able to store soccer players, and soccer team. I also
integrated Masons part please use this version Mason we do not want to
be starting over and over. Vishakh please add the stuff you have to
integrate the frontend and backend to these files. Modify the getEmail
method